+++ Native/CharLS/jlscodecfactory.h
@@ -7,16 +7,16 @@
 
 #include <memory>
 
-struct JlsCustomParameters;
+struct JpegLSPresetCodingParameters;
 struct JlsParameters;
 
 template<typename STRATEGY>
 class JlsCodecFactory
 {
 public:
-    std::unique_ptr<STRATEGY> GetCodec(const JlsParameters& info, const JlsCustomParameters&);
+    std::unique_ptr<STRATEGY> GetCodec(const JlsParameters& params, const JpegLSPresetCodingParameters&);
 private:
-    std::unique_ptr<STRATEGY> GetCodecImpl(const JlsParameters& info);
+    std::unique_ptr<STRATEGY> GetCodecImpl(const JlsParameters& params);
 };
 
 #endif